+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to the "vscode-sticky-selection" extension will be documente
 
 ## [Unreleased]
 
+## [1.2.0]
+
+- Change `sticky-selection.exitStickySelectionMode` to take the interval before exiting sticky-selection-mode after executing the argument command as an argument.
+
 ## [1.1.0]
 
-- Change `exitStickySelectionMode` to take a command to execute before exiting sticky-selection-mode as an argument.
+- Change `sticky-selection.exitStickySelectionMode` to take a command to execute before exiting sticky-selection-mode as an argument.
 
 ## [1.0.0]